added and implemented AuthProvider
site loads and am able to login and logout.
