+++ frontend/src/auth/AuthProvider.js
@@ -1,0 +1,47 @@
+// frontend/src/auth/AuthProvider.js
+
+import React, { createContext, useState, useEffect, useContext } from 'react';
+import { getCurrentUser, isAuthenticated, login as authLogin, logout as authLogout } from './auth';
+
+const AuthContext = createContext();
+
+export const AuthProvider = ({ children }) => {
+  const [isLoggedIn, setIsLoggedIn] = useState(isAuthenticated());
+  const [user, setUser] = useState(null);
+
+  useEffect(() => {
+    const fetchUser = async () => {
+      if (isLoggedIn) {
+        const currentUser = await getCurrentUser();
+        setUser(currentUser);
+      }
+    };
+
+    fetchUser();
+  }, [isLoggedIn]);
+
+  const login = async (email, password) => {
+    try {
+      await authLogin(email, password);
+      setIsLoggedIn(true);
+      const currentUser = await getCurrentUser();
+      setUser(currentUser);
+    } catch (error) {
+      throw error;
+    }
+  };
+
+  const logout = () => {
+    authLogout();
+    setIsLoggedIn(false);
+    setUser(null);
+  };
+
+  return (
+    <AuthContext.Provider value={{ isLoggedIn, user, login, logout }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
+
+export const useAuth = () => useContext(AuthContext);
