+++ frontend/src/index.js
@@ -5,12 +5,14 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
 import 'bulma/css/bulma.min.css';
-
+import { AuthProvider } from './auth/AuthProvider';
 
 ReactDOM.render(
   <React.StrictMode>
     <Router>
-      <App />
+    <AuthProvider>
+        <App />
+      </AuthProvider>
     </Router>
   </React.StrictMode>,
   document.getElementById('root')
